[frontend] Observed Data route padding (#12796)
